redesigned db relationship between user and organization, adding a staff join table. Partially implemented the behavior. Worked on cleaning up some tests. Added a redirect from the devise sign_up page.

+++ db/migrate/20140612211256_drop_programs_users.rb
@@ -1,9 +1,11 @@
 class DropProgramsUsers < ActiveRecord::Migration
   def up
-  	drop_table :programs_users
+    if ActiveRecord::Base.connection.table_exists? :programs_users
+  	   drop_table :programs_users
+    end   
   end
 
-  def down
+  def down    
   	create_table :programs_users, id: false do |t|
     	t.belongs_to :programs
     	t.belongs_to :users
